+++ package.json
@@ -33,8 +33,8 @@
     "prepublish": "npm run build",
     "test:coverage": "cd test && istanbul cover _mocha -- build/index.js",
     "test:integration": "cd test && python integration_test.py",
-    "test:debug": "mocha test/build/index.js --debug-brk",
-    "test": "mocha test/build/index.js"
+    "test:debug": "mocha test/build/test*.js --debug-brk",
+    "test": "mocha test/build/test*.js"
   },
   "repository": {
     "type": "git",
