--- test/src/index.ts
@@ -1,11 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-'use strict';
-
-export * from './testcomm';
-export * from './testconfig';
-export * from './testcontents';
-export * from './testkernel';
-export * from './testmocksocket';
-export * from './testmockxhr';
-export * from './testsession';
